Renamed message => messages to follow naming convention
