+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/messages/UnableToSendMessageException.java
@@ -1,4 +1,4 @@
-package gov.hhs.cdc.trustedintermediary.etor.message;
+package gov.hhs.cdc.trustedintermediary.etor.messages;
 
 public class UnableToSendMessageException extends Exception {
     public UnableToSendMessageException(String message, Throwable cause) {
