+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/results/ResultSender.java
@@ -1,6 +1,6 @@
 package gov.hhs.cdc.trustedintermediary.etor.results;
 
-import gov.hhs.cdc.trustedintermediary.etor.message.UnableToSendMessageException;
+import gov.hhs.cdc.trustedintermediary.etor.messages.UnableToSendMessageException;
 import java.util.Optional;
 
 public interface ResultSender {
